Add OrganizationCardSection component
