+++ src/components/Organization/OrganizationCardSection.css
@@ -1,0 +1,23 @@
+@import "../../theme/colors";
+@import "../../theme/shadows";
+
+.content {
+  display: flex;
+  flex-direction: column;
+  flex-grow: 1;
+  padding: 2em;
+}
+
+.content h4 {
+  margin-top: 1em;
+  margin-bottom: 0.5em;
+}
+
+.section > div {
+  background-color: white;
+  padding: 1em;
+}
+
+.section > div p {
+  margin-bottom: 4px;
+}
