+++ src/components/Organization/__test__/OrganizationCard.test.js
@@ -11,25 +11,6 @@ import user from '../../../fetch/__test__/user.json'
 describe('<OrganizationCard />', () => {
   const { sourceCatalog, producers } = organizationDetail
 
-  describe('At least one error exists', () => {
-    it('should display all errors', () => {
-      const errors = ['une erreur', 'une deuxième erreur']
-      const err = <Errors errors={errors} />
-      const wrapper = shallow(<OrganizationCard organization={user.organizations[0]} metrics={metrics} sourceCatalog={sourceCatalog} producers={producers} errors={errors} />)
-
-      expect(wrapper).to.contain(err)
-    })
-  })
-
-  describe('No error', () => {
-    it('should display metrics', () => {
-      const organizationMetrics = <OrganizationMetrics metrics={metrics} />
-      const wrapper = shallow(<OrganizationCard organization={user.organizations[0]} metrics={metrics} sourceCatalog={sourceCatalog} producers={producers} errors={[]} />)
-
-      expect(wrapper).to.contain(organizationMetrics)
-    })
-  })
-
   describe('Logo', () => {
     it('should display organization logo', () => {
       const wrapper = shallow(<OrganizationCard organization={user.organizations[0]} metrics={metrics} sourceCatalog={sourceCatalog} producers={producers} errors={[]} />)
